style: Change naming convention for orientation derivatives
In all of the literature, "rotation rate" is not the derivative da/dt,
but actually the ω tensor, where a = k exp(- ∫ω dt).
In Fraters 2021 these are distinct tensors,
whereas at the moment we just calculate da/dt in one go.
It may be necessary to split this up for some advection methods.