fix: GBS now carries over previous orientation
Fixes the grain boundary sliding routine to match what is defined in the
original fortran, where dotacs and rot are set to zero for grains that
are sliding, i.e. there is zero rotation _rate_, but not resetting the
whole orientation matrix to the initial one. Now simply carries over the
previous orientation.

Also adds symmetry asserts to the GBS test.