fix: Complete elastic decomposition and M-index and simplify API
Elastic decomposition now provides all symmetry component percentages.

M-index now takes into account the lattice system.
The maximum angle for random input still goes a bit over 120 degrees but
this might just be the tail of the distribution resulting from the
numerics? The mean seems to be slightly low for random input, around 80
instead of 90 degrees, but not too bad. Still, this method needs some
proper tests.

Most postprocessing functions should take a stack of orientations for a
series of texture snapshots, so NxMx3x3 orientations. In some cases
where the diagnostics are cheap to calculate this was not changed, since
it can be trivially achieved with list comprehension.