test: Rework 2D shear GBM test
The 2D shear GBM test is now an ensemble test that compares ensemble
results from pypdrex with ensemble results from the Fortran for the same
parameters. The default number of grains in pydrex was raised again from
2500 to 3500 to be closer to the number originally used in DRex.

Using the 41 runs (`seeds_nearX45`) is enough to pass the checks, and
the test completes in about 10:30 CPU hours using a bit less than 15G
RAM. This can't be run in CI but can be run in less than an hour on
moderately powerful personal PCs (use the custom `--ncpus` arg to pydrex
to run in parallel).