feat: Add builtin mesh generator (2D) and h5part reader
The mesh generator relies on gmsh which is installed as an optional
dependency. `import gmsh` crashes the pdoc github runner though for some
reason so needs to be excluded. The old h5part reader script is removed
and replaced with a first-class CLI tool as well, which makes testing
easier.