test: Use AssertionErrors for failing doctests, not pydrex.Error
Semantically, we should maintain the difference between a test failure
and a test error. Only raise pydrex.Error to intercept
doctes.UnexpectedException which denotes a SyntaxError or similar in a
doctest.

Although pytest offers an inbuilt way to run doctests with
--doctest-modules, I have found that it is not very flexible (there is
no way to filter which doctests are run, or only run doctests and not
the rest of the suite -- such things are possible with this wrapper by
using the pytest -k="<pattern>" command line flag)