dev: Migrate from Makefile to justfile
The [just](https://github.com/casey/just) build command runner is "just"
a better `make`, for our purposes. In particular, it properly terminates
the documentation build if `setuptools-scm` is not installed. I'm sure
there is a way to cast spells at `make` to make it do that, but why?
Whenever there's a big .PHONY line in a Makefile, it means that `make`
is probably the wrong tool. Also, the old Makefile I provided was only
valid for GNU make (make portability is a bit gross).

Also, `just` has binaries built for every major platform so installing
is not a hassle.