More documentation and visualisation fixes (#214)
* docs: Improve documentation and add simple <200LOC cornerflow example

* docs: Add output figure for cornerflow example

* fix: Fix voigt averaging to use new stiffness tensors API

* docs: Add PyPI and GitHub buttons to docs homepage and comment CSS file

* dev: Migrate from Makefile to justfile

The [just](https://github.com/casey/just) build command runner is "just"
a better `make`, for our purposes. In particular, it properly terminates
the documentation build if `setuptools-scm` is not installed. I'm sure
there is a way to cast spells at `make` to make it do that, but why?
Whenever there's a big .PHONY line in a Makefile, it means that `make`
is probably the wrong tool. Also, the old Makefile I provided was only
valid for GNU make (make portability is a bit gross).

Also, `just` has binaries built for every major platform so installing
is not a hassle.

* dev: Fix docs action to use just instead of make

* test: Remove alpha kwarg which is not supported by streamplot

* dev: Use github actions to setup just instead of relying on apt

* fix: Add enstatite CPO component to standalone example

* fix: Typo and wording

---------

Co-authored-by: Thomas Duvernay <td75013@hotmail.fr>